fix(security): replace std::system with shell-free posix_spawn helper
CodeQL flagged 21 cpp/command-line-injection alerts in tools/editor/.
All matched the same pattern: build a shell command string from
argv[0] + a user-supplied path, then std::system() it. Even though
the threat model (user invokes their own CLI on their own machine)
makes the alert mostly academic, the std::system path is also
fragile — paths with spaces, quotes, or shell metacharacters
silently break.

Add tools/editor/cli_subprocess.{hpp,cpp} exposing a single
runChild(argv0, args, quiet=false) that uses posix_spawn on POSIX
and CreateProcess on Windows. No shell, argv passed verbatim,
optional stdout/stderr redirect to /dev/null (NUL on Windows).

Refactor 14 call sites across cli_convert.cpp, cli_data_tree.cpp,
cli_format_validate.cpp, cli_items.cpp, cli_random.cpp,
cli_repair.cpp, cli_spawn_audit.cpp.

Also fix two cpp/integer-multiplication-cast-to-long alerts:
- cli_gen_texture.cpp:3049 — seeds.reserve grid-size product
- cli_convert_single.cpp:224 — vector size for DBC record block
Both now widen one operand to size_t before multiplying.

+++ tools/editor/cli_convert_single.cpp
@@ -221,7 +221,8 @@ int handleConvertJsonDbc(int& i, int argc, char** argv) {
     // strings reuse the same offset in the block. The first byte
     // of the block is always '\0' so offset=0 means empty string,
     // matching Blizzard's convention.
-    std::vector<uint8_t> recordBytes(recordCount * recordSize, 0);
+    std::vector<uint8_t> recordBytes(
+        static_cast<size_t>(recordCount) * static_cast<size_t>(recordSize), 0);
     std::vector<uint8_t> stringBlock;
     stringBlock.push_back(0);  // leading NUL — empty-string offset
     std::unordered_map<std::string, uint32_t> stringOffsets;
